+++ pyproject.toml
@@ -83,12 +83,12 @@ dev = [
 ]
 
 [project.scripts]
-micgeom_app = "spectacoular.apps.micgeom:main"
-msm_app = "spectacoular.apps.measurement:main"
-bf_example_app = "spectacoular.apps.bf_example:main"
-rotating_example_app = "spectacoular.apps.rotating_example:main"
-level_meter_app = "spectacoular.apps.level_meter:main"
-data_viewer_app = "spectacoular.apps.data_viewer:main"
+micgeom_app = "spectacoular.apps.micgeom_app:main"
+msm_app = "spectacoular.apps.measurement_app:main"
+bf_example_app = "spectacoular.apps.bf_example_app:main"
+rotating_example_app = "spectacoular.apps.rotating_example_app:main"
+level_meter_app = "spectacoular.apps.level_meter_app:main"
+data_viewer_app = "spectacoular.apps.data_viewer_app:main"
 
 [project.urls]
 homepage = "https://acoular.github.io/spectacoular/"
