--- src/spectacoular/apps/measurement/config_files/settings_apollo.ini
@@ -1,18 +1,0 @@
-# I've set the INI settings such that it is workflow format from left to right.
-# All values not followed @ implies setting all values.
-# Values followed by @ implies setting values at specific channels. Use "-" between index to indicate range.
-
-# lets say we want to set the Coupling to AC at channel 1-2, 4 and then DC at channel 3.
-# We can set it as Coupling = AC@[1-2 4], DC@[3]
-
-# Note: The ENABLED setting determines which channels are active and will be used for data acquisition.
-# Make sure to set the ENABLED channels according to your required data acquisition needs.
-
-[Device] # sets all settings on the device level
-Mode=HarmEnv_88200 # mainly used for setting the sampling rate which can be set to 88200, 44100, 22050, 11025, ...
-
-[AnalogInput]
-Coupling = AC
-ENABLED = 0,1@[1-2] # activate all channels
-ICP = 0
-sensitivity = 0.016
