--- src/spectacoular/apps/measurement/micgeom/array_64.xml
@@ -1,67 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>									
-<MicArray name="array_64">									
-	<pos Name="Point 	1	" x="	0.152	" y="	0.1141	" z="	0	"/>
-	<pos Name="Point 	2	" x="	0.134	" y="	0.1021	" z="	0	"/>
-	<pos Name="Point 	3	" x="	0.1043	" y="	0.1036	" z="	0	"/>
-	<pos Name="Point 	4	" x="	0.0596	" y="	0.1104	" z="	0	"/>
-	<pos Name="Point 	5	" x="	0.0798	" y="	0.0667	" z="	0	"/>
-	<pos Name="Point 	6	" x="	0.0659	" y="	0.0497	" z="	0	"/>
-	<pos Name="Point 	7	" x="	0.0262	" y="	0.0551	" z="	0	"/>
-	<pos Name="Point 	8	" x="	0.0272	" y="	0.0286	" z="	0	"/>
-	<pos Name="Point 	9	" x="	0	" y="	0	" z="	0	"/>
-	<pos Name="Point 	10	" x="	0.004	" y="	-0.0175	" z="	0	"/>
-	<pos Name="Point 	11	" x="	0.0162	" y="	-0.0078	" z="	0	"/>
-	<pos Name="Point 	12	" x="	0.0162	" y="	0.0078	" z="	0	"/>
-	<pos Name="Point 	13	" x="	0.004	" y="	0.0175	" z="	0	"/>
-	<pos Name="Point 	14	" x="	-0.0112	" y="	0.0141	" z="	0	"/>
-	<pos Name="Point 	15	" x="	-0.018	" y="	0	" z="	0	"/>
-	<pos Name="Point 	16	" x="	-0.0112	" y="	-0.0141	" z="	0	"/>
-	<pos Name="Point 	17	" x="	-0.145	" y="	0.1228	" z="	0	"/>
-	<pos Name="Point 	18	" x="	-0.1294	" y="	0.1079	" z="	0	"/>
-	<pos Name="Point 	19	" x="	-0.1242	" y="	0.0786	" z="	0	"/>
-	<pos Name="Point 	20	" x="	-0.1209	" y="	0.0335	" z="	0	"/>
-	<pos Name="Point 	21	" x="	-0.0828	" y="	0.0629	" z="	0	"/>
-	<pos Name="Point 	22	" x="	-0.0631	" y="	0.0531	" z="	0	"/>
-	<pos Name="Point 	23	" x="	-0.0595	" y="	0.0133	" z="	0	"/>
-	<pos Name="Point 	24	" x="	-0.034	" y="	0.0202	" z="	0	"/>
-	<pos Name="Point 	25	" x="	0.0056	" y="	0.1899	" z="	0	"/>
-	<pos Name="Point 	26	" x="	0.0037	" y="	0.1685	" z="	0	"/>
-	<pos Name="Point 	27	" x="	-0.016	" y="	0.1461	" z="	0	"/>
-	<pos Name="Point 	28	" x="	-0.0492	" y="	0.1155	" z="	0	"/>
-	<pos Name="Point 	29	" x="	-0.0024	" y="	0.104	" z="	0	"/>
-	<pos Name="Point 	30	" x="	0.0022	" y="	0.0825	" z="	0	"/>
-	<pos Name="Point 	31	" x="	-0.0267	" y="	0.0548	" z="	0	"/>
-	<pos Name="Point 	32	" x="	-0.0054	" y="	0.0391	" z="	0	"/>
-	<pos Name="Point 	33	" x="	-0.0874	" y="	-0.1687	" z="	0	"/>
-	<pos Name="Point 	34	" x="	-0.0764	" y="	-0.1502	" z="	0	"/>
-	<pos Name="Point 	35	" x="	-0.049	" y="	-0.1386	" z="	0	"/>
-	<pos Name="Point 	36	" x="	-0.0058	" y="	-0.1254	" z="	0	"/>
-	<pos Name="Point 	37	" x="	-0.0429	" y="	-0.0947	" z="	0	"/>
-	<pos Name="Point 	38	" x="	-0.0378	" y="	-0.0733	" z="	0	"/>
-	<pos Name="Point 	39	" x="	0.0003	" y="	-0.061	" z="	0	"/>
-	<pos Name="Point 	40	" x="	-0.0121	" y="	-0.0376	" z="	0	"/>
-	<pos Name="Point 	41	" x="	-0.1864	" y="	-0.0368	" z="	0	"/>
-	<pos Name="Point 	42	" x="	-0.1651	" y="	-0.0339	" z="	0	"/>
-	<pos Name="Point 	43	" x="	-0.1389	" y="	-0.0481	" z="	0	"/>
-	<pos Name="Point 	44	" x="	-0.1016	" y="	-0.0736	" z="	0	"/>
-	<pos Name="Point 	45	" x="	-0.1008	" y="	-0.0255	" z="	0	"/>
-	<pos Name="Point 	46	" x="	-0.0809	" y="	-0.0162	" z="	0	"/>
-	<pos Name="Point 	47	" x="	-0.0475	" y="	-0.0382	" z="	0	"/>
-	<pos Name="Point 	48	" x="	-0.0369	" y="	-0.014	" z="	0	"/>
-	<pos Name="Point 	49	" x="	0.1839	" y="	-0.0477	" z="	0	"/>
-	<pos Name="Point 	50	" x="	0.1634	" y="	-0.0411	" z="	0	"/>
-	<pos Name="Point 	51	" x="	0.146	" y="	-0.0169	" z="	0	"/>
-	<pos Name="Point 	52	" x="	0.1235	" y="	0.0223	" z="	0	"/>
-	<pos Name="Point 	53	" x="	0.1019	" y="	-0.0208	" z="	0	"/>
-	<pos Name="Point 	54	" x="	0.0799	" y="	-0.0205	" z="	0	"/>
-	<pos Name="Point 	55	" x="	0.0594	" y="	0.0138	" z="	0	"/>
-	<pos Name="Point 	56	" x="	0.0393	" y="	-0.0034	" z="	0	"/>
-	<pos Name="Point 	57	" x="	0.0774	" y="	-0.1735	" z="	0	"/>
-	<pos Name="Point 	58	" x="	0.0697	" y="	-0.1534	" z="	0	"/>
-	<pos Name="Point 	59	" x="	0.0778	" y="	-0.1247	" z="	0	"/>
-	<pos Name="Point 	60	" x="	0.0944	" y="	-0.0827	" z="	0	"/>
-	<pos Name="Point 	61	" x="	0.0473	" y="	-0.0926	" z="	0	"/>
-	<pos Name="Point 	62	" x="	0.0338	" y="	-0.0753	" z="	0	"/>
-	<pos Name="Point 	63	" x="	0.0478	" y="	-0.0378	" z="	0	"/>
-	<pos Name="Point 	64	" x="	0.0218	" y="	-0.0329	" z="	0	"/>
-</MicArray>
