--- src/spectacoular/apps/measurement/micgeom/transit_vogel_array75_2021.xml
@@ -1,78 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<MicArray name="transit_vogel_array75_2021">
-	<pos Name="Point 0" x="0.6621849262965267" y="1.4224715502235739" z="0.0"/>
-	<pos Name="Point 1" x="0.21718228484109453" y="0.9339935110008345" z="0.0"/>
-	<pos Name="Point 2" x="0.8461055965360964" y="0.954588474980619" z="0.0"/>
-	<pos Name="Point 3" x="0.9385839217935604" y="0.6616121419826659" z="0.0"/>
-	<pos Name="Point 4" x="0.9253065416230684" y="0.42136480527424686" z="0.0"/>
-	<pos Name="Point 5" x="0.5386475070358162" y="0.5026323803658033" z="0.0"/>
-	<pos Name="Point 6" x="0.44070840279707" y="0.7132411475568626" z="0.0"/>
-	<pos Name="Point 7" x="0.18180606233614455" y="0.4848877049290592" z="0.0"/>
-	<pos Name="Point 8" x="0.5090038350949851" y="0.30391608024244393" z="0.0"/>
-	<pos Name="Point 9" x="0.8097419927009011" y="0.22447280336143374" z="0.0"/>
-	<pos Name="Point 10" x="0.5846666086361735" y="0.07712238827394872" z="0.0"/>
-	<pos Name="Point 11" x="0.3042996436264548" y="0.11614231482232712" z="0.0"/>
-	<pos Name="Point 12" x="0.22007441474578118" y="0.2821717448461561" z="0.0"/>
-	<pos Name="Point 13" x="0.02941" y="0.13492612838201787" z="0.0"/>
-	<pos Name="Point 14" x="0.0294" y="0.659512723540244" z="0.0"/>
-	<pos Name="Point 15" x="1.0542452118054715" y="0.05240523648984963" z="0.0"/>
-	<pos Name="Point 16" x="-0.15410534669475298" y="1.1863669097874225" z="0.0"/>
-	<pos Name="Point 17" x="-0.28941014671737364" y="0.8086395479715535" z="0.0"/>
-	<pos Name="Point 18" x="-0.7093793764422573" y="0.9331504468080317" z="0.0"/>
-	<pos Name="Point 19" x="-0.6410971287054623" y="0.6112604783613221" z="0.0"/>
-	<pos Name="Point 20" x="-0.3160449922585062" y="0.5272113541877872" z="0.0"/>
-	<pos Name="Point 21" x="-0.08326533984878193" y="0.387" z="0.0"/>
-	<pos Name="Point 22" x="-0.1980737838061575" y="0.22882741059035286" z="0.0"/>
-	<pos Name="Point 23" x="-0.4888581260414106" y="0.34559" z="0.0"/>
-	<pos Name="Point 24" x="-0.8200519182810656" y="0.3844" z="0.0"/>
-	<pos Name="Point 25" x="-1.0502094858133668" y="0.6394156559903045" z="0.0"/>
-	<pos Name="Point 26" x="-1.185048789886645" y="0.3369700420430796" z="0.0"/>
-	<pos Name="Point 27" x="-0.49576507699534245" y="0.17238359742471035" z="0.0"/>
-	<pos Name="Point 28" x="-0.8297217782452505" y="0.1588223266086388" z="0.0"/>
-	<pos Name="Point 29" x="-0.3165383868344601" y="0.03359471018379763" z="0.0"/>
-	<pos Name="Point 30" x="-1.135876731677525" y="0.06316368011180302" z="0.0"/>
-	<pos Name="Point 31" x="-0.6945305087150998" y="0.02941" z="0.0"/>
-	<pos Name="Point 32" x="-0.11730168289264097" y="-0.06447477021353683" z="0.0"/>
-	<pos Name="Point 33" x="-0.10292318990270416" y="-0.2297214817672568" z="0.0"/>
-	<pos Name="Point 34" x="-0.453667408236504" y="-0.11236038672730839" z="0.0"/>
-	<pos Name="Point 35" x="-0.3586365301974249" y="-0.3060741355883632" z="0.0"/>
-	<pos Name="Point 36" x="-0.03421401730543418" y="-0.44394580107541337" z="0.0"/>
-	<pos Name="Point 37" x="-0.6622566993069511" y="-0.2764685771967733" z="0.0"/>
-	<pos Name="Point 38" x="-0.35245139873171916" y="-0.50529875946615" z="0.0"/>
-	<pos Name="Point 39" x="-0.9430676757860224" y="-0.14915376766960473" z="0.0"/>
-	<pos Name="Point 40" x="-0.7323949459297668" y="-0.47501848568820937" z="0.0"/>
-	<pos Name="Point 41" x="-0.21166757316167523" y="-0.6981073047106694" z="0.0"/>
-	<pos Name="Point 42" x="-1.1087967709748754" y="-0.3414097177226247" z="0.0"/>
-	<pos Name="Point 43" x="-0.6855488962525367" y="-0.7018605921771978" z="0.0"/>
-	<pos Name="Point 44" x="-1.1918646587881683" y="-0.5908667603728902" z="0.0"/>
-	<pos Name="Point 45" x="-0.5185401541855765" y="-0.959044467683244" z="0.0"/>
-	<pos Name="Point 46" x="-1.1974380844442105" y="-0.9184097685923367" z="0.0"/>
-	<pos Name="Point 47" x="-0.21099950992527883" y="-1.2970983830860374" z="0.0"/>
-	<pos Name="Point 48" x="0.7780022307765977" y="-0.06280745986826942" z="0.0"/>
-	<pos Name="Point 49" x="0.43308760397547685" y="-0.0948977390149789" z="0.0"/>
-	<pos Name="Point 50" x="0.16879638674981975" y="-0.1320989975470696" z="0.0"/>
-	<pos Name="Point 51" x="0.16358166893057655" y="-0.31291448242255654" z="0.0"/>
-	<pos Name="Point 52" x="0.5136560755143763" y="-0.2597459305424995" z="0.0"/>
-	<pos Name="Point 53" x="0.8449739037069989" y="-0.2664567614549272" z="0.0"/>
-	<pos Name="Point 54" x="1.1424391942156806" y="-0.4869211741969742" z="0.0"/>
-	<pos Name="Point 55" x="0.7524712526970603" y="-0.49769383853977994" z="0.0"/>
-	<pos Name="Point 56" x="0.4224617911518962" y="-0.4423512934403869" z="0.0"/>
-	<pos Name="Point 57" x="0.17218644478093595" y="-0.6011421877906161" z="0.0"/>
-	<pos Name="Point 58" x="0.4861959776491003" y="-0.7165969898569656" z="0.0"/>
-	<pos Name="Point 59" x="0.9062284650578707" y="-0.7896744848040675" z="0.0"/>
-	<pos Name="Point 60" x="0.441" y="-1.0660430968490464" z="0.0"/>
-	<pos Name="Point 61" x="0.05364206900115028" y="-0.882574556886036" z="0.0"/>
-	<pos Name="Point 62" x="1.1655606948858848" y="-1.3792643776023625" z="0.0"/>
-	<pos Name="Point 63" x="1.1813557426491834" y="-0.1942563789408871" z="0.0"/>
-	<pos Name="Point 64" x="-1.567596036770909" y="0.6112766130074717" z="0.0"/>
-	<pos Name="Point 65" x="-1.5833830684531778" y="0.2185996464489045" z="0.0"/>
-	<pos Name="Point 66" x="-1.6685098804494716" y="-0.35457117646045894" z="0.0"/>
-	<pos Name="Point 67" x="-1.4135954452384205" y="-0.10570420674180803" z="0.0"/>
-	<pos Name="Point 68" x="-2.070572395333195" y="0.012121423932940741" z="0.0"/>
-	<pos Name="Point 69" x="2.256726082040413" y="-0.2531579497827827" z="0.0"/>
-	<pos Name="Point 70" x="1.5189234737845274" y="-0.046598147024314966" z="0.0"/>
-	<pos Name="Point 71" x="1.6011138202258568" y="-0.40756452204455434" z="0.0"/>
-	<pos Name="Point 72" x="1.4375394025486217" y="0.488671414584856" z="0.0"/>
-	<pos Name="Point 73" x="1.2743142522720718" y="0.23573082219303665" z="0.0"/>
-	<pos Name="Point 74" x="1.8796226420526543" y="0.18815895489685355" z="0.0"/>
-</MicArray>
