+++ src/spectacoular/apps/measurement_app/config_files/settings_typhoon.ini
@@ -1,0 +1,15 @@
+# I've set the INI settings such that it is workflow format from left to right.
+# All values not followed @ implies setting all values.
+# Values followed by @ implies setting values at specific channels. Use "-" between index to indicate range.
+
+# lets say we want to set the Coupling to AC at channel 1-2, 4 and then DC at channel 3.
+# We can set it as Coupling = AC@[1-2 4], DC@[3]
+
+# Note: The ENABLED setting determines which channels are active and will be used for data acquisition.
+# Make sure to set the ENABLED channels according to your required data acquisition needs.
+
+[AnalogInput]
+Coupling = AC
+ENABLED = 0, 1@[1-64] # only activate the first 64 channels
+ICP = 0 # disable ICP (activate if needed for specific channels, e.g. ICP = 2mA@[1-64])
+sensitivity = 0.016
