+++ src/spectacoular/apps/measurement_app/micgeom/minidsp_uma16.xml
@@ -1,0 +1,18 @@
+<?xml version="1.0" encoding="utf-8"?><MicArray name="uma16">
+   <pos Name="Point 1"	x="	0.021000000"	y="	-0.063000000"	z="0" />
+   <pos Name="Point 2"	x="	0.063000000"	y="	-0.063000000"	z="0" />
+   <pos Name="Point 3"	x="	0.021000000"	y="	-0.021000000"	z="0" />
+   <pos Name="Point 4"	x="	0.063000000"	y="	-0.021000000"	z="0" />
+   <pos Name="Point 5"	x="	0.021000000"	y="	0.021000000"	z="0" />
+   <pos Name="Point 6"	x="	0.063000000"	y="	0.021000000"	z="0" />
+   <pos Name="Point 7"	x="	0.021000000"	y="	0.063000000"	z="0" />
+   <pos Name="Point 8"	x="	0.063000000"	y="	0.063000000"	z="0" />
+   <pos Name="Point 9"	x="	-0.063000000"	y="	0.063000000"	z="0" />
+   <pos Name="Point 10"	x="	-0.021000000"	y="	0.063000000"	z="0" />
+   <pos Name="Point 11"	x="	-0.063000000"	y="	0.021000000"	z="0" />
+   <pos Name="Point 12"	x="	-0.021000000"	y="	0.021000000"	z="0" />
+   <pos Name="Point 13"	x="	-0.063000000"	y="	-0.021000000"	z="0" />
+   <pos Name="Point 14"	x="	-0.021000000"	y="	-0.021000000"	z="0" />
+   <pos Name="Point 15"	x="	-0.063000000"	y="	-0.063000000"	z="0" />
+   <pos Name="Point 16"	x="	-0.021000000"	y="	-0.063000000"	z="0" />
+</MicArray>
