+++ src/spectacoular/apps/measurement_app/micgeom/tub_vogel64.xml
@@ -1,0 +1,67 @@
+<?xml version="1.0" encoding="utf-8"?>
+<MicArray name="tub_vogel64">
+   <pos Name="Point 1"	x="	0.171669218"	y="	0.017056542"	z="0" />
+   <pos Name="Point 2"	x="	0.271377821"	y="	0.130816725"	z="0" />
+   <pos Name="Point 3"	x="	0.429344538"	y="	0.101786815"	z="0" />
+   <pos Name="Point 4"	x="	0.553153354"	y="	0.009057487"	z="0" />
+   <pos Name="Point 5"	x="	0.705086017"	y="	0.106648562"	z="0" />
+   <pos Name="Point 6"	x="	0.574969965"	y="	0.220712658"	z="0" />
+   <pos Name="Point 7"	x="	0.427502308"	y="	0.281878052"	z="0" />
+   <pos Name="Point 8"	x="	0.273547015"	y="	0.281278267"	z="0" />
+   <pos Name="Point 9"	x="	0.120631127"	y="	0.194538437"	z="0" />
+   <pos Name="Point 10"	x="	0.068756808"	y="	0.330836809"	z="0" />
+   <pos Name="Point 11"	x="	0.191672135"	y="	0.427696178"	z="0" />
+   <pos Name="Point 12"	x="	0.343589500"	y="	0.463680679"	z="0" />
+   <pos Name="Point 13"	x="	0.509769377"	y="	0.444110111"	z="0" />
+   <pos Name="Point 14"	x="	0.355053278"	y="	0.643826591"	z="0" />
+   <pos Name="Point 15"	x="	0.181139028"	y="	0.612437328"	z="0" />
+   <pos Name="Point 16"	x="	0.038182314"	y="	0.535628685"	z="0" />
+   <pos Name="Point 17"	x="	-0.043673404"	y="	0.696962065"	z="0" />
+   <pos Name="Point 18"	x="	-0.165366510"	y="	0.577284174"	z="0" />
+   <pos Name="Point 19"	x="	-0.389074494"	y="	0.534596476"	z="0" />
+   <pos Name="Point 20"	x="	-0.598075395"	y="	0.401755115"	z="0" />
+   <pos Name="Point 21"	x="	-0.419998313"	y="	0.372291254"	z="0" />
+   <pos Name="Point 22"	x="	-0.235321236"	y="	0.435512797"	z="0" />
+   <pos Name="Point 23"	x="	-0.061108242"	y="	0.417777463"	z="0" />
+   <pos Name="Point 24"	x="	-0.097099346"	y="	0.256685465"	z="0" />
+   <pos Name="Point 25"	x="	-0.242585685"	y="	0.281208722"	z="0" />
+   <pos Name="Point 26"	x="	-0.393447191"	y="	0.219485527"	z="0" />
+   <pos Name="Point 27"	x="	-0.580151385"	y="	0.228434494"	z="0" />
+   <pos Name="Point 28"	x="	-0.683303191"	y="	0.017056586"	z="0" />
+   <pos Name="Point 29"	x="	-0.513914961"	y="	0.082251103"	z="0" />
+   <pos Name="Point 30"	x="	-0.300569263"	y="	0.090495454"	z="0" />
+   <pos Name="Point 31"	x="	-0.154560347"	y="	0.115702168"	z="0" />
+   <pos Name="Point 32"	x="	-0.001399365"	y="	0.121736583"	z="0" />
+   <pos Name="Point 33"	x="	-0.099547740"	y="	-0.111206581"	z="0" />
+   <pos Name="Point 34"	x="	-0.235813336"	y="	-0.066307107"	z="0" />
+   <pos Name="Point 35"	x="	-0.401839457"	y="	-0.023212872"	z="0" />
+   <pos Name="Point 36"	x="	-0.574267609"	y="	-0.111305455"	z="0" />
+   <pos Name="Point 37"	x="	-0.703695465"	y="	-0.237259960"	z="0" />
+   <pos Name="Point 38"	x="	-0.553717462"	y="	-0.333118348"	z="0" />
+   <pos Name="Point 39"	x="	-0.438400780"	y="	-0.189424296"	z="0" />
+   <pos Name="Point 40"	x="	-0.283477425"	y="	-0.204204406"	z="0" />
+   <pos Name="Point 41"	x="	-0.106870314"	y="	-0.267580289"	z="0" />
+   <pos Name="Point 42"	x="	-0.011395002"	y="	-0.381819782"	z="0" />
+   <pos Name="Point 43"	x="	-0.122755334"	y="	-0.488403121"	z="0" />
+   <pos Name="Point 44"	x="	-0.242532107"	y="	-0.357269315"	z="0" />
+   <pos Name="Point 45"	x="	-0.395886535"	y="	-0.374768446"	z="0" />
+   <pos Name="Point 46"	x="	-0.443231652"	y="	-0.549173046"	z="0" />
+   <pos Name="Point 47"	x="	-0.271453344"	y="	-0.544271359"	z="0" />
+   <pos Name="Point 48"	x="	-0.075060577"	y="	-0.664422194"	z="0" />
+   <pos Name="Point 49"	x="	0.059039977"	y="	-0.566135744"	z="0" />
+   <pos Name="Point 50"	x="	0.171355056"	y="	-0.707396516"	z="0" />
+   <pos Name="Point 51"	x="	0.276812998"	y="	-0.567155417"	z="0" />
+   <pos Name="Point 52"	x="	0.497668007"	y="	-0.479287390"	z="0" />
+   <pos Name="Point 53"	x="	0.685902211"	y="	-0.303377910"	z="0" />
+   <pos Name="Point 54"	x="	0.505281410"	y="	-0.309916967"	z="0" />
+   <pos Name="Point 55"	x="	0.328544863"	y="	-0.414300549"	z="0" />
+   <pos Name="Point 56"	x="	0.144734426"	y="	-0.436290234"	z="0" />
+   <pos Name="Point 57"	x="	0.166738845"	y="	-0.280243254"	z="0" />
+   <pos Name="Point 58"	x="	0.319695484"	y="	-0.260608445"	z="0" />
+   <pos Name="Point 59"	x="	0.459510432"	y="	-0.159335517"	z="0" />
+   <pos Name="Point 60"	x="	0.640672413"	y="	-0.129891894"	z="0" />
+   <pos Name="Point 61"	x="	0.358038647"	y="	-0.042217957"	z="0" />
+   <pos Name="Point 62"	x="	0.232174256"	y="	-0.117213870"	z="0" />
+   <pos Name="Point 63"	x="	0.039269818"	y="	-0.208035744"	z="0" />
+   <pos Name="Point 64"	x="	0.058818183"	y="	-0.062742833"	z="0" />
+</MicArray>
